+++ app/src/main/java/nodomain/freeyourgadget/gadgetbridge/GBDevice.java
@@ -1,0 +1,23 @@
+package nodomain.freeyourgadget.gadgetbridge;
+
+public class GBDevice {
+    private final String name;
+    private final String address;
+
+    public GBDevice(String address, String name) {
+        this.address = address;
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public String getStatus() {
+        return "";
+    }
+}